Initial Commit (v1).
downloadalbum is a module that provides a way of downloading a complete album. A ZIP archive is generated on-the-fly in a efficient way: it doesn't burn your CPU nor eat your RAM. There is nothing stored on disk so it will not waste space.

The module add a floppy icon on an album when the visitor has the view_full right.

This module is based on downloadfullsize.